foodgram/requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.6
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.1.0
drf-extra-fields==3.4.0
flake8==5.0.0
idna==3.3
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
oauthlib==3.2.0
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.9.0
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.11
